requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.118.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
multidict==6.7.0
propcache==0.4.0
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.22.0
